tentatively removed a fix, see whether spears extend/stack infinitely with autofire again reworked snow ench, it is now stardust ench lite (scuffed rn but w/e)

+++ FargoSoulsUtil.cs
@@ -61,7 +61,7 @@ namespace FargowiltasSouls
             return (int)(modifier * npcDamage / inherentHostileProjMultiplier / Math.Max(npcDamageCalculationsOffset, worldDamage));
         }
 
-        public static void AllCritEquals(Player player, int crit)
+        public static void AllCritEquals(Player player, float crit)
         {
             player.GetCritChance(DamageClass.Generic) = crit;
